Scalar graphql Float => scala Double by default.

+++ packages/graphql-codegen-scala-common/src/scalars.ts
@@ -5,5 +5,5 @@ export const DEFAULT_SCALARS: NormalizedScalarsMap = {
   String: "String",
   Boolean: "Boolean",
   Int: "Int",
-  Float: "Float"
+  Float: "Double"
 };
